remove participant round1
