+++ frost-client/src/participant/mod.rs
@@ -3,5 +3,3 @@ pub mod cli;
 pub mod comms;
 
 pub mod input;
-pub mod round1;
-pub mod round2;
